Update increaser_decreaser_calculations.R
I do not know what etiquette  is when  collaborating on code, so I left yours alone and added a section of mine which filters the metaData and uses it as a lookup table for the summary stats.  I created variables for each of the sites based upon their SWAMP codes (above, below) and made a pivot table to view the results.